+++ lib/src/events/solar_events.dart
@@ -1,0 +1,125 @@
+import 'package:viet_date_time/src/models/event.dart';
+import 'package:viet_date_time/src/models/event_list.dart';
+
+final getSolarEvents = VietEventList(events: {
+  DateTime(0, 1, 1): [
+    VietEvent(date: DateTime(0, 1, 1), title: 'Tết Dương Lịch'),
+  ],
+  DateTime(0, 2, 14): [
+    VietEvent(date: DateTime(0, 2, 14), title: 'Lễ tình nhân (Valentine)'),
+  ],
+  DateTime(0, 2, 27): [
+    VietEvent(date: DateTime(0, 2, 27), title: 'Ngày thầy thuốc Việt Nam'),
+  ],
+  DateTime(0, 3, 8): [
+    VietEvent(date: DateTime(0, 3, 8), title: 'Ngày Quốc tế Phụ nữ'),
+  ],
+  DateTime(0, 3, 23): [
+    VietEvent(date: DateTime(0, 3, 23), title: 'Ngày Quốc tế Hạnh phúc'),
+  ],
+  DateTime(0, 3, 26): [
+    VietEvent(
+      date: DateTime(0, 3, 26),
+      title: 'Ngày thành lập Đoàn TNCS Hồ Chí Minh',
+    ),
+  ],
+  DateTime(0, 4, 1): [
+    VietEvent(date: DateTime(0, 4, 1), title: 'Ngày Cá tháng Tư'),
+  ],
+  DateTime(0, 4, 30): [
+    VietEvent(date: DateTime(0, 4, 30), title: 'Ngày giải phóng miền Nam'),
+  ],
+  DateTime(0, 5, 1): [
+    VietEvent(date: DateTime(0, 5, 1), title: 'Ngày Quốc tế Lao động'),
+  ],
+  DateTime(0, 5, 7): [
+    VietEvent(date: DateTime(0, 5, 7), title: 'Ngày chiến thắng Điện Biên Phủ'),
+  ],
+  DateTime(0, 5, 13): [
+    VietEvent(date: DateTime(0, 5, 13), title: 'Ngày của mẹ'),
+  ],
+  DateTime(0, 5, 19): [
+    VietEvent(
+        date: DateTime(0, 5, 19), title: 'Ngày sinh chủ tịch Hồ Chí Minh'),
+  ],
+  DateTime(0, 6, 1): [
+    VietEvent(date: DateTime(0, 6, 1), title: 'Ngày Quốc tế thiếu nhi'),
+  ],
+  DateTime(0, 6, 17): [
+    VietEvent(date: DateTime(0, 6, 17), title: 'Ngày của cha'),
+  ],
+  DateTime(0, 6, 21): [
+    VietEvent(date: DateTime(0, 6, 21), title: 'Ngày báo chí Việt Nam'),
+  ],
+  DateTime(0, 6, 28): [
+    VietEvent(date: DateTime(0, 6, 28), title: 'Ngày gia đình Việt Nam'),
+  ],
+  DateTime(0, 7, 11): [
+    VietEvent(date: DateTime(0, 7, 11), title: 'Ngày dân số thế giới'),
+  ],
+  DateTime(0, 7, 27): [
+    VietEvent(date: DateTime(0, 7, 27), title: 'Ngày Thương binh liệt sĩ'),
+  ],
+  DateTime(0, 7, 28): [
+    VietEvent(
+        date: DateTime(0, 7, 28), title: 'Ngày thành lập Công đoàn Việt Nam'),
+  ],
+  DateTime(0, 8, 19): [
+    VietEvent(date: DateTime(0, 8, 19), title: 'Ngày tổng khởi nghĩa'),
+  ],
+  DateTime(0, 9, 2): [
+    VietEvent(date: DateTime(0, 9, 2), title: 'Ngày Quốc Khánh'),
+  ],
+  DateTime(0, 9, 10): [
+    VietEvent(
+      date: DateTime(0, 9, 10),
+      title: 'Ngày thành lập Mặt trận Tổ quốc Việt Nam',
+    ),
+  ],
+  DateTime(0, 10, 1): [
+    VietEvent(date: DateTime(0, 10, 1), title: 'Ngày Quốc tế Người cao tuổi'),
+  ],
+  DateTime(0, 10, 10): [
+    VietEvent(date: DateTime(0, 10, 10), title: 'Ngày giải phóng thủ đô'),
+  ],
+  DateTime(0, 10, 13): [
+    VietEvent(date: DateTime(0, 10, 13), title: 'Ngày doanh nhân Việt Nam'),
+  ],
+  DateTime(0, 10, 20): [
+    VietEvent(date: DateTime(0, 10, 20), title: 'Ngày Phụ nữ Việt Nam'),
+  ],
+  DateTime(0, 10, 31): [
+    VietEvent(date: DateTime(0, 10, 31), title: 'Ngày Halloween'),
+  ],
+  DateTime(0, 11, 9): [
+    VietEvent(date: DateTime(0, 11, 9), title: 'Ngày pháp luật Việt Nam'),
+  ],
+  DateTime(0, 11, 19): [
+    VietEvent(date: DateTime(0, 11, 19), title: 'Ngày Quốc tế nam giới'),
+  ],
+  DateTime(0, 11, 20): [
+    VietEvent(date: DateTime(0, 11, 20), title: 'Ngày Nhà giáo Việt Nam'),
+  ],
+  DateTime(0, 11, 23): [
+    VietEvent(
+      date: DateTime(0, 11, 23),
+      title: 'Ngày thành lập Hội chữ thập đỏ Việt Nam',
+    ),
+  ],
+  DateTime(0, 12, 1): [
+    VietEvent(
+        date: DateTime(0, 12, 1), title: 'Ngày Thế giới phòng chống AIDS'),
+  ],
+  DateTime(0, 12, 19): [
+    VietEvent(date: DateTime(0, 12, 19), title: 'Ngày Toàn quốc kháng chiến'),
+  ],
+  DateTime(0, 12, 22): [
+    VietEvent(
+      date: DateTime(0, 12, 22),
+      title: 'Ngày thành lập Quân đội nhân dân Việt Nam',
+    ),
+  ],
+  DateTime(0, 12, 24): [
+    VietEvent(date: DateTime(0, 12, 24), title: 'Ngày Lễ Giáng sinh'),
+  ],
+});