+++ lib/src/extension.dart
@@ -1,0 +1,6 @@
+import 'package:viet_date_time/src/viet_date_time.dart';
+
+/// Convert from solar (normal DateTime) to VietDateTime.
+extension SolarToLunarConverter on DateTime {
+  VietDateTime get toVietDateTime => VietDateTime.fromDateTime(this);
+}