+++ lib/viet_date_time.dart
@@ -1,3 +1,5 @@
 library viet_date_time;
 
+export 'src/models/event.dart';
+export 'src/models/event_list.dart';
 export 'src/viet_date_time.dart';