+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.0.1-rc.3
+
+[BREAKING CHANGE]
+
+* Rename from `.toSolar` to `.toDateTime`.
+* Rename from `.fromSolar` to `.fromDateTime`.
+* Rename from `judianDayNumber` to `julianDayNumber`.
+* Add `VietDateTime.lunarEvents` and `VietDateTime.solarEvents`.
+* Add extension `.toVietDateTime` to convert from `DateTime` to `VietDateTime`.
+
 ## 0.0.1-rc.2
 
 * Use positioned parameters instead of named.